1. 配环境产生的env修改(并不是当前env) 2. 环境不同导致的一些小的硬编码区别, 比如numpy.float->float; logger的xls->xlsx否则报错

+++ standalone_eval_v2.py
@@ -153,8 +153,8 @@ class ScanNetEval(object):
         dist_confs = [self.distance_confs[0]]
 
         # results: class x iou
-        ap = np.zeros((len(dist_threshes), len(self.eval_class_labels), len(ious)), np.float)
-        rc = np.zeros((len(dist_threshes), len(self.eval_class_labels), len(ious)), np.float)
+        ap = np.zeros((len(dist_threshes), len(self.eval_class_labels), len(ious)), float)
+        rc = np.zeros((len(dist_threshes), len(self.eval_class_labels), len(ious)), float)
         for di, (min_region_size, distance_thresh, distance_conf) in enumerate(
             zip(min_region_sizes, dist_threshes, dist_confs)
         ):
